pyproject.toml
[tool.poetry]
name = "hw14"
version = "0.1.0"
description = ""
authors = ["AndriiRohovyk <77483613+AndreyRogovik@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.1.1"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
libgravatar = "^1.0.4"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.39.0"
passlib = "^1.7.4"
fastapi-mail = "^1.4.1"
python-jose = "^3.3.0"
uvicorn = "^0.28.0"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
httpx = "^0.27.0"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
